TEST: version_less_than() must compare releases if version same
If the version parts of two x.y.z-r version strings are the same,
version_less_than() must compare the release parts rather than defaulting
to whatever the last return value happened to be.

This can be tested with this:

	sh ./vercmp.sh 2.6.32-589.el6 2.6.32-592.el6.ppc64

Reversing the parameters should flip the output indication rather than
leaving it the same.

Signed-off-by: David Howells <dhowells@redhat.com>