man: Squash multiline NAME entries
The norm for man pages is that the NAME line is formatted as a single
line without line breaks. The patched pages violate that norm,
and break various tools that process man page input
(such as my own scripts to format pages for man7.org).

Squash the NAME lines in the pages that have this problem, so that
a single line is produced. This necessarily entails abbreviating
the descriptive text somewhat, but I don't think this is a loss.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>