keyctl: use keyctl_read_alloc() in dump_key_tree_aux()
dump_key_tree_aux() (part of 'keyctl show') was racy: it allocated a
buffer for the keyring contents, then read the keyring.  But it's
possible that keys are added to the keyring concurrently.  This is
problematic for two reasons.  First, when keyctl_read() is passed a
buffer that is too small, it is unspecified whether it is filled or not.
Second, even if the buffer is filled, some keys (not necessarily even
the newest ones) would be omitted from the listing.

Switch to keyctl_read_alloc() which handles the "buffer too small" case
correctly by retrying the read.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: David Howells <dhowells@redhat.com>